Created test parameters package and moved relevant tests into it
