+++ src/test/java/uk/co/ramp/covid/simulation/parameters/CovidParametersTest.java
@@ -1,4 +1,4 @@
-package uk.co.ramp.covid.simulation.covid;
+package uk.co.ramp.covid.simulation.parameters;
 
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.parameters.CovidParameters;
