+++ src/test/java/uk/co/ramp/covid/simulation/parameters/PopulationDistributionTest.java
@@ -1,8 +1,10 @@
-package uk.co.ramp.covid.simulation.population;
+package uk.co.ramp.covid.simulation.parameters;
 
+import org.junit.Assert;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.parameters.PopulationDistribution;
 import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
+import uk.co.ramp.covid.simulation.population.Person;
 import uk.co.ramp.covid.simulation.util.InvalidParametersException;
 import uk.co.ramp.covid.simulation.testutil.SimulationTest;
 
@@ -59,7 +61,7 @@ public class PopulationDistributionTest extends SimulationTest {
 
         PopulationDistribution.SexAge s = dist.sample();
         assertEquals(30, s.getAge());
-        assertEquals(Person.Sex.MALE, s.getSex());
+        Assert.assertEquals(Person.Sex.MALE, s.getSex());
     }
 
     @Test
